src/main/webapp/app/entities/office/delete/office-delete-dialog.component.ts
import { Component } from '@angular/core';
import { FormsModule } from '@angular/forms';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';
import { IOffice } from '../office.model';
import { OfficeService } from '../service/office.service';

@Component({
  standalone: true,
  templateUrl: './office-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class OfficeDeleteDialogComponent {
  office?: IOffice;

  constructor(
    protected officeService: OfficeService,
    protected activeModal: NgbActiveModal,
  ) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.officeService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
